Adapt app to syncronizer (#719)
* Start of the synchronizer, paths, and sorting

* Add tests for the unique files in the loadout

* WIP

* Tons of test debugging, need to think about this a bit

* Can apply changes!

* Sanity tests pass

* Few more tests

* Can get the new Disk State (during ingest)

* Reset the subbed game folders when we start the next test

* Can ingest into a filetree

* Can ingest into a flattened loadout

* WIP need to update the loadout from the flattened loadout

* Supports deleting and adding files

* Tests work but likely need work

* Can merge loadouts

* Fixup some broken tests

* Run datamodel tests serially to reduce errors

* Fix locator on Linux

* Fix path sanitation issue on Linux

* Fix all the bad paths

* Fix those tests I broke

* Bit of code cleanup

* Cleanup and WIP with generated files

* Generated files are properly created

* Add more comments, code cleanup

* Update ingestion code to properly handle generated files

* Fix broken test

* Rearrange parameters for better consistency

* Handle PR feedback

* Fix comment on Mod.cs

* Start of adapting the app to the new synchronizer

* Cyberpunk tests pass

* WIP

* CanGeneratePlugins file fails 20% of the time :|

* Fix a bug with the plugin sorter

* WIP code fixes

* more test fixes, WIP

* WIP

* Few more bug fixes

* Tons of small fixes and cleanups

* Tons of small fixes

* Tests pass, can it really be?

* Remove unused file

* More comments

* Rename the file to get it to work on Linux

* Make sure we backup new files during ingestion

* Code cleanup

* Fix the stack overflow bug in AVmTest.cs

* Handle some feedback from the PR

* Fix broken tests

* Handle PR feedback

+++ tests/NexusMods.StandardGameLocators.TestHelpers/StubbedGames/StubbedGame.cs
@@ -40,7 +40,7 @@ public class StubbedGame : AGame, IEADesktopGame, IEpicGame, IOriginGame, ISteam
     private bool _initalized;
 
     public StubbedGame(ILogger<StubbedGame> logger, IEnumerable<IGameLocator> locators,
-        IFileSystem fileSystem, IServiceProvider provider) : base(locators)
+        IFileSystem fileSystem, IServiceProvider provider) : base(provider)
     {
         _serviceProvider = provider;
         _logger = logger;
@@ -58,7 +58,8 @@ public class StubbedGame : AGame, IEADesktopGame, IEpicGame, IOriginGame, ISteam
         {
             foreach (var result in locator.Find(this))
             {
-                result.Path.DeleteDirectory(true);
+                if (result.Path.DirectoryExists())
+                    result.Path.DeleteDirectory(true);
                 var locations = new Dictionary<LocationId, AbsolutePath>
                 {
                     [LocationId.Game] = EnsureFiles(result.Path, LocationId.Game),
